+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ### Fixes
 - Fix issue where some images would accidentally use default background color
   in some terminals instead of the desired color.
+- Properly round volume.
 
 ## v0.5.12
 ### New Features